pyproject.toml
[project]
name = "listserial"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Unkn0wnN4m3", email = "87839850+Unkn0wnN4m3@users.noreply.github.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pyserial==3.5",
]

[project.scripts]
listserial = "listserial:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
